backend/requirements.txt
fastapi==0.129.1
uvicorn==0.41.0
python-dotenv==1.1.0
anthropic==0.83.0
python-multipart>=0.0.18
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
python-telegram-bot[job-scheduler]==21.10
